convert src to javascript
